register.jsx  added
